start moving things to rank.py
