+++ mne/utils/__init__.py
@@ -33,7 +33,7 @@ from .testing import (_memory_usage, run_tests_if_main, requires_sklearn,
                       requires_numpydoc, requires_tvtk, requires_freesurfer,
                       requires_nitime, requires_fs_or_nibabel, requires_dipy,
                       requires_neuromag2ft)
-from .numerics import (hashfunc, md5sum, estimate_rank, _compute_row_norms,
+from .numerics import (hashfunc, md5sum, _compute_row_norms,
                        _reg_pinv, random_permutation, _reject_data_segments,
                        compute_corr, _get_inst_data, array_split_idx,
                        sum_squared, split_list, _gen_events, create_slices,
