+++ mne/utils/numerics.py
@@ -73,78 +73,6 @@ def sum_squared(X):
     return np.dot(X_flat, X_flat)
 
 
-def estimate_rank(data, tol='auto', return_singular=False, norm=True):
-    """Estimate the rank of data.
-
-    This function will normalize the rows of the data (typically
-    channels or vertices) such that non-zero singular values
-    should be close to one.
-
-    Parameters
-    ----------
-    data : array
-        Data to estimate the rank of (should be 2-dimensional).
-    tol : float | 'auto'
-        Tolerance for singular values to consider non-zero in
-        calculating the rank. The singular values are calculated
-        in this method such that independent data are expected to
-        have singular value around one. Can be 'auto' to use the
-        same thresholding as ``scipy.linalg.orth``.
-    return_singular : bool
-        If True, also return the singular values that were used
-        to determine the rank.
-    norm : bool
-        If True, data will be scaled by their estimated row-wise norm.
-        Else data are assumed to be scaled. Defaults to True.
-
-    Returns
-    -------
-    rank : int
-        Estimated rank of the data.
-    s : array
-        If return_singular is True, the singular values that were
-        thresholded to determine the rank are also returned.
-    """
-    data = data.copy()  # operate on a copy
-    if norm is True:
-        norms = _compute_row_norms(data)
-        data /= norms[:, np.newaxis]
-    s = linalg.svd(data, compute_uv=False, overwrite_a=True)
-    rank = _estimate_rank_from_s(s, tol)
-    if return_singular is True:
-        return rank, s
-    else:
-        return rank
-
-
-def _estimate_rank_from_s(s, tol='auto'):
-    """Estimate the rank of a matrix from its singular values.
-
-    Parameters
-    ----------
-    s : list of float
-        The singular values of the matrix.
-    tol : float | 'auto'
-        Tolerance for singular values to consider non-zero in calculating the
-        rank. Can be 'auto' to use the same thresholding as
-        ``scipy.linalg.orth``.
-
-    Returns
-    -------
-    rank : int
-        The estimated rank.
-    """
-    if isinstance(tol, str):
-        if tol != 'auto':
-            raise ValueError('tol must be "auto" or float')
-        eps = np.finfo(float).eps
-        tol = len(s) * np.amax(s) * eps
-
-    tol = float(tol)
-    rank = np.sum(s > tol)
-    return rank
-
-
 def _compute_row_norms(data):
     """Compute scaling based on estimated norm."""
     norms = np.sqrt(np.sum(data ** 2, axis=1))
